start central FoamAdapter.hpp, move readers/foamFields.hpp  -> readers.hpp
