+++ include/FoamAdapter/FoamAdapter.hpp
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// SPDX-FileCopyrightText: 2024 NeoFOAM authors
+
+#include "NeoFOAM/finiteVolume/cellCentred/interpolation/surfaceInterpolation.hpp"
+#include "NeoFOAM/finiteVolume/cellCentred/operators/gaussGreenDiv.hpp"
+
+#include "FoamAdapter/meshAdapter.hpp"
+#include "FoamAdapter/readers.hpp"
+#include "FoamAdapter/writers.hpp"
+#include "FoamAdapter/setup.hpp"
