+++ include/FoamAdapter/meshAdapter.hpp
@@ -13,7 +13,7 @@
 #include "NeoFOAM/core/executor/executor.hpp"
 #include "NeoFOAM/mesh/unstructured.hpp"
 
-#include "readers/foamFields.hpp"
+#include "readers.hpp"
 
 namespace Foam
 {
